requirements.txt
annotated-types==0.7.0
anyio==4.9.0
axios==0.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docx==0.2.4
fastapi==0.115.12
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.27
llmwhisperer-client==2.3.0
lxml==5.3.2
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.74.0
orjson==3.10.16
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
unstract==0.0.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zstandard==0.23.0
